+++ extensions/iceworks-material-helper/webpack.config.js
@@ -1,12 +1,17 @@
 const path = require('path');
+const CopyPlugin = require('copy-webpack-plugin');
 
 const tsConfigPath = path.join(__dirname, 'tsconfig.json');
+const destPath = path.resolve(__dirname, 'build');
 
 const config = {
   target: 'node',
   entry: './src/extension.ts',
+  node: {
+    __dirname: false,
+  },
   output: {
-    path: path.resolve(__dirname, 'build'),
+    path: destPath,
     filename: 'extension.js',
     libraryTarget: 'commonjs2',
     devtoolModuleFilenameTemplate: '../[resource-path]',
@@ -35,6 +40,22 @@ const config = {
       },
     ],
   },
+  plugins: [
+    new CopyPlugin({
+      patterns: [
+        {
+          from: 'src/templates/*.ejs',
+          globOptions: {
+            gitignore: true,
+          },
+          to() {
+            return `${destPath }/[name].[ext]`;
+          },
+        },
+      ],
+      options: { concurrency: 10 },
+    }),
+  ],
 };
 
 module.exports = (env, argv) => {