+++ extensions/iceworks-time-master/src/extension.ts
@@ -1,5 +1,5 @@
 import { ExtensionContext, commands } from 'vscode';
-import { Recorder, recordDAU } from '@iceworks/recorder';
+import { recordDAU } from '@iceworks/recorder';
 import { createTimerTreeView, TimerProvider, createTimerStatusBar, autoSetEnableViewsConfig } from './views';
 import { openFileInEditor } from './utils/common';
 import { getInterface as getKeystrokeStats } from './recorders/keystrokeStats';
@@ -7,20 +7,16 @@ import { getInterface as getUsageStatsRecorder } from './recorders/usageStats';
 import { activate as activateWalkClock, deactivate as deactivateWalkClock } from './managers/walkClock';
 import { generateProjectSummaryReport, generateUserSummaryReport } from './managers/data';
 import logger from './utils/logger';
-
-// eslint-disable-next-line
-const { name, version } = require('../package.json');
-const recorder = new Recorder(name, version);
+import recorder from './utils/recorder';
 
 const keystrokeStatsRecorder = getKeystrokeStats();
 const usageStatsRecorder = getUsageStatsRecorder();
 
 export async function activate(context: ExtensionContext) {
-  logger.debug('[TimeMaster][extension] activate!');
   const { subscriptions, globalState } = context;
 
-  // do not wait for async, let subsequent views be created
-  activateWalkClock();
+  console.log('Congratulations, your extension "iceworks-time-master" is now active!');
+  recorder.recordActivate();
 
   autoSetEnableViewsConfig(globalState);
 
@@ -32,13 +28,6 @@ export async function activate(context: ExtensionContext) {
   const timerStatusBar = await createTimerStatusBar();
   timerStatusBar.activate();
 
-  keystrokeStatsRecorder.activate().catch((e) => {
-    logger.error('[TimeMaster][extension] activate keystrokeStatsRecorder got error:', e);
-  });
-  usageStatsRecorder.activate().catch((e) => {
-    logger.error('[TimeMaster][extension] activate usageStatsRecorder got error:', e);
-  });
-
   subscriptions.push(
     commands.registerCommand('iceworks-time-master.openFileInEditor', (fsPath: string) => {
       openFileInEditor(fsPath);
@@ -47,9 +36,6 @@ export async function activate(context: ExtensionContext) {
         module: 'command',
         action: 'openFileInEditor',
       });
-    }),
-    commands.registerCommand('iceworks-time-master.sendKeystrokeStatsMap', () => {
-      keystrokeStatsRecorder.sendData();
     }),
     commands.registerCommand('iceworks-time-master.refreshTimerTree', () => {
       timerProvider.refresh();
@@ -82,17 +68,42 @@ export async function activate(context: ExtensionContext) {
       });
     }),
   );
+
+  // do not wait for async, let subsequent views be created
+  activateWalkClock().catch((e) => {
+    logger.error('[TimeMaster][extension] activate walkClock got error:', e);
+  });
+
+  keystrokeStatsRecorder.activate().catch((e) => {
+    logger.error('[TimeMaster][extension] activate keystrokeStatsRecorder got error:', e);
+  });
+  usageStatsRecorder.activate().catch((e) => {
+    logger.error('[TimeMaster][extension] activate usageStatsRecorder got error:', e);
+  });
 }
 
-export function deactivate() {
-  logger.debug('[TimeMaster][extension] deactivate!');
+export async function deactivate() {
+  logger.disable('console');
+  logger.info('[TimeMaster][extension][deactivate] start!');
 
-  keystrokeStatsRecorder.deactivate().catch((e) => {
-    logger.error('[TimeMaster][extension] deactivate keystrokeStatsRecorder got error:', e);
-  });
-  usageStatsRecorder.deactivate().catch((e) => {
-    logger.error('[TimeMaster][extension] deactivate usageStatsRecorder got error:', e);
-  });
+  try {
+    await Promise.all([
+      keystrokeStatsRecorder.deactivate(),
+      usageStatsRecorder.deactivate(),
+    ]);
+  } catch (e) {
+    logger.error('[TimeMaster][extension][deactivate] recorders got error:', e);
+  }
 
-  deactivateWalkClock();
+  logger.info('[TimeMaster][extension][deactivate] recorders success!');
+
+  try {
+    await deactivateWalkClock();
+  } catch (e) {
+    logger.error('[TimeMaster][extension][deactivate] walkClock got error:', e);
+  }
+
+  logger.info('[TimeMaster][extension][deactivate] walkClock success!');
+
+  logger.info('[TimeMaster][extension][deactivate] done!');
 }